requirements.txt
black==24.8.0
flask==3.0.3
flake8==7.1.1
flake8-bugbear==24.8.19
flake8-import-order==0.18.2
gensim==4.3.3
neo4j==5.24.0
nltk==3.9.1
numpy==1.24.4
openai==1.47.0
python-dotenv==1.0.1
relik==1.0.7
scikit-learn==1.5.2
spacy==3.7.5
torch==2.2.2
transformers==4.41.0
